Converting sea_floor_depth_below_sea_level to float32 to avoid problems for old versions of numpy

+++ opendrift/models/basemodel.py
@@ -728,7 +728,7 @@ class OpenDriftSimulation(PhysicsMethods):
                 self.get_environment(['sea_floor_depth_below_sea_level'],
                                       t, kwargs['lon'], kwargs['lat'],
                                       0.*ones, None)
-            kwargs['z'] = -env['sea_floor_depth_below_sea_level']
+            kwargs['z'] = -env['sea_floor_depth_below_sea_level'].astype('float32')
 
         elements = self.ElementType(**kwargs)
